# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FoodMind Backend API
  version: v1
  description: Public API contract for FoodMind clients.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    CorrelationId:
      description: Accepted or generated request correlation identifier.
      schema:
        type: string
        maxLength: 128
      example: postman-correlation-test
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Optional client-generated correlation identifier.
      schema:
        type: string
        maxLength: 128
      example: postman-correlation-test
...
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Bounded page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    ApiFieldError:
      type: object
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          example: price
        code:
          type: string
          example: POSITIVE_OR_ZERO
        message:
          type: string
          example: Price must be zero or greater.
    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
        - path
        - traceId
        - fieldErrors
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-07-28T10:15:30Z"
        status:
          type: integer
          example: 400
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - MALFORMED_JSON
            - AUTHENTICATION_REQUIRED
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_CONFLICT
            - RATE_LIMITED
            - UPSTREAM_UNAVAILABLE
            - INTERNAL_ERROR
          example: VALIDATION_ERROR
        message:
          type: string
          example: The request contains invalid fields.
        path:
          type: string
          example: /api/v1/food-records
        traceId:
          type: string
          example: trace-123
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    PageResponse:
      type: object
      required:
        - items
        - page
        - size
        - totalItems
        - totalPages
        - hasNext
      properties:
        items:
          type: array
          items: {}
        page:
          type: integer
          minimum: 0
          example: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        totalItems:
          type: integer
          format: int64
          minimum: 0
          example: 0
        totalPages:
          type: integer
          minimum: 0
          example: 0
        hasNext:
          type: boolean
          example: false
...
    UserRecipeRequest:
      type: object
      required: [name, servings, ingredients, steps]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        servings:
          type: integer
          minimum: 1
          maximum: 50
        imageUrl:
          type: string
          maxLength: 2048
          nullable: true
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 80
        allergenHints:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 80
        ingredients:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 500
        steps:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 1000
    UserRecipeResponse:
      type: object
      required: [id, name, servings, tags, allergenHints, ingredients, steps, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        servings:
          type: integer
          minimum: 1
          maximum: 50
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items: { type: string }
        allergenHints:
          type: array
          items: { type: string }
        ingredients:
          type: array
          items: { type: string }
        steps:
          type: array
          items: { type: string }
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 0
    UserRecipePageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/UserRecipeResponse'
...
  responses:
    BadRequest:
      description: Request validation or JSON parsing failed.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or invalid.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    NotFound:
      description: Resource was not found or cannot be disclosed.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with current resource state.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
paths:
...
  /recipes:
    get:
      tags:
        - Recipes
      operationId: listUserRecipes
      summary: List the authenticated user's recipes.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Owner-scoped recipe page.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipePageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Recipes
      operationId: createUserRecipe
      summary: Create an owned recipe.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRecipeRequest'
      responses:
        '201':
          description: Created recipe.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            ETag:
              description: Current optimistic-lock version.
              schema:
                type: string
                example: '"0"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /recipes/{id}:
    get:
      tags:
        - Recipes
      operationId: getUserRecipe
      summary: Read an owned recipe.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Owned recipe detail.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            ETag:
              description: Current optimistic-lock version.
              schema:
                type: string
                example: '"0"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Recipes
      operationId: updateUserRecipe
      summary: Update an owned recipe with optimistic locking.
      description: Requires If-Match with the current numeric version. Non-owners receive a non-disclosing 404.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
            example: '"0"'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRecipeRequest'
      responses:
        '200':
          description: Updated recipe.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            ETag:
              description: New optimistic-lock version.
              schema:
                type: string
                example: '"1"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - Recipes
      operationId: deleteUserRecipe
      summary: Soft-delete an owned recipe.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recipe deleted or already absent for this owner.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...